Tests added
The DBManagerTest file is not working, not yet sure why. In the meantime, Test.py includes simple if/else + print tests for the methods. This tells me that the problem is in the usage of the unittest module, not in the logic in DBManager.
DBManager was supposed to include functions for events and notifications too, but I will get the unittest sorted before adding them.